+++ TinkoffASDKUI/TinkoffASDKUITests/TestCases/TinkoffPayControllerTests.swift
@@ -5,6 +5,7 @@
 //  Created by Никита Васильев on 31.05.2023.
 //
 
+@testable import TinkoffASDKCore
 @testable import TinkoffASDKUI
 import XCTest
 
@@ -73,6 +74,243 @@ final class TinkoffPayControllerTests: BaseTestCase {
 
         // then
         XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithTimeoutCallsCount, 1)
+    }
+
+    func test_thatControllerNotifiesDelegate_whenStatusIsSuccess() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = true
+        repeatedRequestHelperMock.executeWithWaitingIfNeededActionShouldExecute = true
+        paymentStatusServiceMock.getPaymentStateCompletionClosureInputs = [.success(.fake(status: .authorized))]
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerDidOpenURLCallsCount, 1)
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithSuccessfulCallsCount, 1)
+    }
+
+    func test_thatControllerNotifiesDelegate_whenStatusIsFailed() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = true
+        repeatedRequestHelperMock.executeWithWaitingIfNeededActionShouldExecute = true
+        paymentStatusServiceMock.getPaymentStateCompletionClosureInputs = [.success(.fake(status: .rejected))]
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithFailedCallsCount, 1)
+    }
+
+    func test_thatControllerRetriesRequest_whenStatusIsSuccess() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 2)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = true
+        repeatedRequestHelperMock.executeWithWaitingIfNeededActionShouldExecute = true
+        paymentStatusServiceMock.getPaymentStateCompletionClosureInputs = [.success(.fake(status: .checked3ds))]
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCallsCount, 1)
+    }
+
+    func test_thatControllerCompletesWithTimeout_whenStatusIsSuccess() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = true
+        repeatedRequestHelperMock.executeWithWaitingIfNeededActionShouldExecute = true
+        paymentStatusServiceMock.getPaymentStateCompletionClosureInputs = [.success(.fake(status: .checked3ds))]
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithTimeoutCallsCount, 1)
+    }
+
+    func test_thatControllerRetriesRequest_whenStatusIsFailed() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 2)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = true
+        repeatedRequestHelperMock.executeWithWaitingIfNeededActionShouldExecute = true
+        paymentStatusServiceMock.getPaymentStateCompletionClosureInputs = [.failure(ErrorStub())]
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(repeatedRequestHelperMock.executeWithWaitingIfNeededCallsCount, 2)
+    }
+
+    func test_thatControllerCompletesWithError_whenStatusIsFailed() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = true
+        repeatedRequestHelperMock.executeWithWaitingIfNeededActionShouldExecute = true
+        paymentStatusServiceMock.getPaymentStateCompletionClosureInputs = [.failure(ErrorStub())]
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithErrorCallsCount, 1)
+    }
+
+    func test_thatControllerNotifiesDelegate_whenErrorOccurred() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        paymentService.initPaymentCompletionInput = .failure(ErrorStub())
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithErrorCallsCount, 1)
+    }
+
+    func test_thatControllerNotifiesDelegate_whenGetLinkErrorOccurred() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .failure(ErrorStub())
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerCompletedWithErrorCallsCount, 1)
+    }
+
+    func test_thatControllerNotifiesDelegate_whenUnableToOpenExternalApp() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        queueMock.asyncWorkShouldExecute = true
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+        tinkoffPayServiceMock.stubbedGetTinkoffPayLinkCompletion = .success(.fake())
+        applicationOpenerMock.openCompletionClosureInput = false
+
+        // when
+        sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+
+        // then
+        XCTAssertEqual(delegateMock.tinkoffPayControllerInabilityToOpenTinkoffPayCallsCount, 1)
+    }
+
+    func test_thatControllerCancelsRequest_whenStatusIsSuccess() {
+        // given
+        let sut = prepareSut(paymentStatusRetriesCount: 1)
+
+        paymentService.initPaymentCompletionInput = .success(.fake())
+
+        // when
+        let process = sut.performPayment(paymentFlow: .full(paymentOptions: .fake()), method: .fake())
+        process.cancel()
+
+        // then
+        XCTAssertEqual((process as? TinkoffPayController.Process)?.isActive, false)
+    }
+
+    func test_errorLocalization_whenCouldNotOpenTinkoffPayApp() {
+        // given
+        let url = URL(string: "https://vk.com/")!
+
+        // when
+        let tpayError = TinkoffPayController.Error.couldNotOpenTinkoffPayApp(url: url)
+
+        // then
+        XCTAssertEqual(tpayError.errorDescription, "Could not open TinkoffPay App with url \(url)")
+        XCTAssertEqual(
+            tpayError.recoverySuggestion,
+            url.scheme.map { "For Tinkoff Pay to work correctly, add the scheme \($0) to the list of LSApplicationQueriesSchemes at info.plist" }
+        )
+    }
+
+    func test_errorLocalization_whenDidNotWaitForSuccessfulPaymentState() {
+        // given
+        let payload = GetPaymentStatePayload.fake()
+        let error = ErrorStub()
+
+        // when
+        let tpayError = TinkoffPayController.Error.didNotWaitForSuccessfulPaymentState(
+            lastReceivedPaymentState: payload,
+            underlyingError: error
+        )
+
+        // then
+        XCTAssertEqual(
+            tpayError.errorDescription,
+            "Something went wrong in the payment process: the payment did not reach final status completed. Last received payment payload: \(payload). Underlying error: \(error)"
+        )
+        XCTAssertNil(tpayError.recoverySuggestion)
+    }
+
+    func test_errorLocalization_whenDidNotWaitForSuccessfulPaymentStateAndValuesAreNil() {
+        // when
+        let tpayError = TinkoffPayController.Error.didNotWaitForSuccessfulPaymentState(
+            lastReceivedPaymentState: nil,
+            underlyingError: nil
+        )
+
+        // then
+        XCTAssertEqual(
+            tpayError.errorDescription,
+            "Something went wrong in the payment process: the payment did not reach final status completed"
+        )
+        XCTAssertNil(tpayError.recoverySuggestion)
+    }
+
+    func test_errorLocalization_whenDidReceiveFailedPaymentState() {
+        // given
+        let payload = GetPaymentStatePayload.fake()
+
+        // when
+        let tpayError = TinkoffPayController.Error.didReceiveFailedPaymentState(payload)
+
+        // then
+        XCTAssertEqual(
+            tpayError.errorDescription,
+            "Something went wrong in the payment process: the payment was rejected. Payload: \(payload)"
+        )
+        XCTAssertNil(tpayError.recoverySuggestion)
     }
 
     // MARK: - Private